feat: add additional metrics to support elastic pool (sqlserver plugin) (#9841)
